Moving KG Embedding Calculations from sun_models to utils/convenience
Moving the functions from exogenous/sun_models to utils/convenience for
a better location:
- calculate_entity_range
- calculate_entity_centroid
- get_entity_embeddings_from_indices
Modifying affected files accordingly:
- mlm_training.py
- nav_supervised_training.py
- nav_training.py
- multihopkg/rl/graph_search/pn.py
